removed all unneccessary files and reorganized vitals files into a vitals folder in components, also renamed things for consistency
