+++ frontend/src/components/vitals/VitalCard.tsx
