--- frontend/src/pages/homepage/Heropage.css
@@ -1,242 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  /* border: solid yellow; */
-}
-
-/* --- MAIN PAGE BACKGROUND --- */
-.hero-page-container {
-  width: 100%;
-  height: 100%;
-  min-height: 100vh;
-  /* background-color: black; */
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  overflow: hidden;
-}
-
-.hero-content-container {
-  display: grid;
-  justify-content: center;
-  align-items: center;
-  grid-template-columns: 1fr 1fr;
-  grid-template-rows: auto;
-  margin: 0 auto;
-  padding: 10rem 0 4rem 0;
-  column-gap: 50px;
-  height: 100%;
-  width: 100%;
-  background-color: transparent !important;
-  position: relative;
-  z-index: 10;
-}
-
-.left-section {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  margin: auto 50px;
-  padding: 100px 0 0 10px;
-  position: relative;
-  bottom: 100px;
-  top: -220px;
-}
-
-.header-hook {
-  display: flex;
-  justify-content: flex-start;
-  font-family: "Aspekta Variable", sans-serif;
-  font-variation-settings: "wght" 600;
-  font-weight: 600;
-  font-size: 40px;
-  line-height: 90px;
-  letter-spacing: -2px;
-  color: white;
-  margin-bottom: 1rem;
-  text-align: left;
-  background: linear-gradient(#ced8cd, #a9cec0 100%, #66c9a8);
-  -webkit-background-clip: text;
-  background-clip: text;
-  -webkit-text-fill-color: transparent;
-  display: inline-block;
-}
-
-.header-hook-dark {
-  background: linear-gradient(#34d399, #10b981);
-  -webkit-background-clip: text;
-  background-clip: text;
-  -webkit-text-fill-color: transparent;
-  display: inline-block;
-}
-
-.hero-description,
-.current-session-banner,
-.session-error {
-  font-family: "Inter", sans-serif;
-  font-size: 18px;
-  font-weight: 400;
-  letter-spacing: -0.2px;
-  color: #f5f5f5;
-  margin-bottom: 3rem;
-  line-height: 24px;
-  max-width: 600px;
-}
-
-.current-session-banner {
-  margin-top: 3rem;
-}
-
-.session-error {
-  color: #f50000;
-}
-
-.session-list {
-  list-style: none;
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  width: 100%;
-}
-
-.session-row {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  margin-bottom: 10px;
-}
-
-.new-session-input,
-.old-session-input {
-  flex: 1;
-  max-width: 260px;
-  height: 50px;
-  margin-left: 20px;
-  padding: 0 16px;
-  border-radius: 12px;
-  border: 1px solid #333;
-  background-color: #181c1a;
-  color: #f5f5f5;
-  font-family: "Inter", sans-serif;
-  font-size: 1rem;
-  outline: none;
-}
-
-.create-session,
-.join-session {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  max-width: 500px;
-  height: 70px;
-  border-radius: 12px;
-  cursor: pointer;
-  text-decoration: none;
-  font-family: "Poppins", sans-serif;
-  font-size: 1.2rem;
-  font-weight: 600;
-  transition: transform 0.2s ease;
-}
-
-.create-session {
-  position: relative;
-  overflow: hidden;
-  color: #000; /* Keeping text black for contrast on the bright green */
-
-  /* Layer 1: Radial glow from the bottom | Layer 2: Surface linear tint */
-  background-image:
-    radial-gradient(circle at 50% 100%, #065f46, #047857 51%),
-    linear-gradient(#10b98122, #065f46);
-
-  background-origin: border-box;
-  background-clip: padding-box, border-box;
-  border: 1px solid transparent;
-
-  /* The bottom "rim" shadow for 3D effect */
-  box-shadow: 0 2px 0 1px #064e3b;
-
-  transition: all 0.2s ease;
-}
-
-.button-light-effect {
-  filter: blur(20px);
-  background-color: #ffffff20; /* Reduced opacity for a subtler look on dark green */
-  border-radius: 50%;
-  width: 80%;
-  height: 100%;
-  position: absolute;
-  inset: -30px auto auto;
-  pointer-events: none;
-}
-
-.join-session {
-  border: 1px solid #333;
-  color: white;
-  background-color: #181c1a;
-}
-
-.create-session:hover,
-.join-session:hover {
-  transform: translateY(-2px); /* Subtle hover lift */
-}
-
-.monitor-section {
-  position: relative;
-  top: -150px;
-  margin: auto;
-  padding: 50px;
-}
-
-.monitor-wrapper {
-  position: relative;
-}
-
-.monitor-glow {
-  position: absolute;
-  inset: 0;
-  background: linear-gradient(
-    to right,
-    rgba(16, 185, 129, 0.15),
-    rgba(6, 182, 212, 0.15)
-  );
-  border-radius: 32px;
-  filter: blur(40px);
-  transform: scale(1.08);
-  z-index: 0;
-  animation: pulse-glow 4s infinite ease-in-out;
-}
-.monitor-box {
-  position: relative;
-  z-index: 1;
-  background: linear-gradient(135deg, #0a0a0f 0%, #12121a 100%);
-  border-radius: 24px;
-  border: 1px solid rgba(16, 185, 129, 0.3);
-  min-height: 600px;
-  min-width: 600px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  box-shadow:
-    0 0 0 1px rgba(6, 182, 212, 0.2),
-    0 25px 50px -12px rgba(0, 0, 0, 0.5);
-}
-
-.preview-label {
-  position: absolute;
-  bottom: -18px;
-  left: 50%;
-  transform: translateX(-50%);
-  background-color: #0a0a0f;
-  border: 1px solid #1e293b;
-  color: #64748b;
-  font-size: 11px;
-  font-family: "Inter", sans-serif;
-  text-transform: uppercase;
-  letter-spacing: 1px;
-  padding: 4px 14px;
-  border-radius: 20px;
-  z-index: 10;
-}
