+++ frontend/src/pages/monitor/MonitorView.css
