--- frontend/src/pages/monitor/monitorTrack.tsx
@@ -1,1 +1,0 @@
-export {};
