+++ frontend/tsconfig.json
@@ -15,6 +15,7 @@
     "isolatedModules": true,
     "noEmit": true,
     "jsx": "react-jsx",
+    "types": ["node"],
 
     // Tailwind configuring
     "paths": {
