+++ frontend/vite.config.ts
@@ -17,7 +17,7 @@ export default defineConfig({
   test: {
     environment: "jsdom",
     globals: true,
-    setupFiles: "./src/setupTests.js",
+    setupFiles: "./src/setupTests.ts",
   },
   resolve: {
     alias: { "@": path.resolve(__dirname, "./src") },
